+++ crates/wdl-engine/src/stdlib/matches.rs
@@ -1,6 +1,5 @@
 //! Implements the `matches` function from the WDL standard library.
 
-use regex::Regex;
 use wdl_analysis::types::PrimitiveType;
 use wdl_ast::Diagnostic;
 
@@ -29,7 +28,7 @@ fn matches(context: CallContext<'_>) -> Result<Value, Diagnostic> {
         .coerce_argument(1, PrimitiveType::String)
         .unwrap_string();
 
-    let regex = Regex::new(pattern.as_str())
+    let regex = super::cached_regex(pattern.as_str())
         .map_err(|e| function_call_failed(FUNCTION_NAME, &e, context.arguments[1].span))?;
     Ok(regex.is_match(input.as_str()).into())
 }