+++ crates/wdl-engine/src/stdlib/split.rs
@@ -1,6 +1,5 @@
 //! Implements the `split` function from the WDL standard library.
 
-use regex::Regex;
 use wdl_analysis::stdlib::STDLIB as ANALYSIS_STDLIB;
 use wdl_analysis::types::PrimitiveType;
 use wdl_ast::Diagnostic;
@@ -33,7 +32,7 @@ fn split(context: CallContext<'_>) -> Result<Value, Diagnostic> {
         .coerce_argument(1, PrimitiveType::String)
         .unwrap_string();
 
-    let regex = Regex::new(delimiter.as_str())
+    let regex = super::cached_regex(delimiter.as_str())
         .map_err(|e| function_call_failed(FUNCTION_NAME, &e, context.arguments[1].span))?;
 
     let elements = regex